requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
dateparser==1.4.0
distro==1.9.0
Flask==3.1.1
fonttools==4.59.2
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
multidict==6.7.1
narwhals==2.2.0
np==1.0.2
numpy==2.2.4
openai==2.28.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==21.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.6.1
pyjokes==0.8.3
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python_binance==1.0.36
pyttsx3==2.99
pytz==2025.2
pywin32==311
referencing==0.36.2
regex==2026.2.28
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.49.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
websockets==16.0
Werkzeug==3.1.3
yarl==1.23.0
